drivers: adc: rework Nordic SAADC driver
SHIM Nordic modification for the ADC driver controlling the SAADC
peripheral. Replaced HAL based implementation in favor of nrfx driver.

As a next step, it is planned to implement a feature that will allow
the peripheral SAADC timer to be used for sampling, and for this it is
necessary to use the nrfx driver in this SHIM.
This will allow more accurate and faster sampling than the kernel
mechanism currently provides.

(cherry picked from commit f3d250742377ac24af0bc4bf78dd61afaefc2479)

Original-Signed-off-by: Jakub Zymelka <jakub.zymelka@nordicsemi.no>
GitOrigin-RevId: f3d250742377ac24af0bc4bf78dd61afaefc2479
Cr-Build-Id: 8711229266906812305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8711229266906812305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4c23f2bb1b609cf0f0fb653ab8c6b9c97777c2f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6665360
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>