led: fix up the ENOSYS condition to pass coverity
The current ENOSYS check code can proceed for a driver that
hypothetically implements the on() but not the off() function of
vice versa, which would result in a null pointer dereference.

This would be a weird use case but Coverity catches the situation, no
harm in changing the ENOSYS check to fail if the on/off is half
implemented, so let's change it to do that.

(cherry picked from commit 033672114b60a7d7ebec14161bdee3db81af8d3e)

Original-Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
GitOrigin-RevId: 033672114b60a7d7ebec14161bdee3db81af8d3e
Cr-Build-Id: 8713426245664562529
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713426245664562529
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I064c57e65a7af0473d5b89f5678ab898d8d5496b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6608190
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>