tests: spi_loopback: Handle overflow of print latency
The test that checks latency also prints a measurement of the latency
which is just a visual information for the test output. However, this
print uses a subtraction which can have an overflow in the case where
the kernel timer is behind, so just clamp to 0 in that case where there
is an overflow. This change does not affect the test pass/fail logic.

(cherry picked from commit c10c13b961aa8668d325b8c1ee4ef00dcb3e65ea)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: c10c13b961aa8668d325b8c1ee4ef00dcb3e65ea
Cr-Build-Id: 8713380946827786369
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8713380946827786369
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I86cd4a9c8345d2e88b14d5147cdf8b8d8f7bf07b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6608921
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>