drivers: audio: dmic_mcux: remove arbitrary mapping of pdm to dmic channel
The dmic_mcux driver required a mapping of paired dmics to
specify the same pdm number, and would arbitrarily assign the left
channel of the pair to even dmic channel number and the right
to an odd dmic channel number.

Change this so that the pdm number in the mapping is
used as the dmic channel number, and paired dmics are checked to
specify consecutive pdm numbers (instead of the
same pdm number).

This allows users to control explicitly which dmic
channel to use and whether that dmic channel is left or right,
without this arbitrary indirect mapping. This is important
in case they want a dmic that is wired to be right channel
to be assigned to dmic channel 0, which is the only channel
that supports hwvad.

(cherry picked from commit 1f69b91e909cc37301ac15d7e7f30e87c413ab21)

Original-Signed-off-by: Mike J. Chen <mjchen@google.com>
GitOrigin-RevId: 1f69b91e909cc37301ac15d7e7f30e87c413ab21
Cr-Build-Id: 8708669904892160353
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708669904892160353
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib220009043cf4085071b524b9201b8cd216c6dfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6776832
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>