ipm: remove the Intel Audio DSP CAVS host IPM driver
This IPM driver is not being used by SOF and is there simply for
a test which does not provide much additional value compared to
the existing host IPC tests. That IPM specific test has been
removed so there is no need to keep this driver in the tree.
This is in preparation to rewrite the host IPC driver to utilize
the IPC API. There is no need to maintain another driver that is
not being used in SOF.

(cherry picked from commit 0b3b6f30a3dd1c22905a7755410f8c55edb77d13)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: 0b3b6f30a3dd1c22905a7755410f8c55edb77d13
Cr-Build-Id: 8702373416049541713
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8702373416049541713
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I3cf207211d4f05c11c5b876ef1af1df8e09fb459
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6996102
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>