drivers: clock_control: nrf2: align with hw binding names
Currently there is a mismatch between the naming of the hardware and
the drivers targetting the hardware. nrf2_ is used instead of
the actual bindings names, like nrf2_audiopll instead of
nrfs_audiopll. This makes it hard to map drivers to the hardware
they are targetting.

There is historical reason for some of this, namely the same binding
name was used for different hardware, which is why nrf2_ was used
on newer platforms. This is no longer the case though, so drivers
and configs can be named according to the hardware without conflict.

(cherry picked from commit 979a5652890687d6ba1fe0408a3cfb3db8760f70)

Original-Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>
GitOrigin-RevId: 979a5652890687d6ba1fe0408a3cfb3db8760f70
Cr-Build-Id: 8712180537715307889
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8712180537715307889
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic0ba120375ee4238662b061b571b3fe439fdf656
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6643319
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Ting Shen <phoenixshen@chromium.org>