arch: xtensa: Add support for lazy HiFi ctx switching
When lazy HiFi context switching is enabled, the system starts with
the HiFi coprocessor disabled. Should the thread use that coprocessor,
it will generate an exception which in turn will enable the coprocessor
and save/restore the HiFi registers as appropriate. When switching
to a new thread, the HiFi coprocessor is again disabled.

For simplicity, there are no restrictions as to which thread is allowed
to use the coprocessor.

(cherry picked from commit e50851d47641de1489726907ec3e384991f64106)

Original-Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>
GitOrigin-RevId: e50851d47641de1489726907ec3e384991f64106
Cr-Build-Id: 8708805799059677889
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708805799059677889
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6f470c6a133bbed2773ed328f3126c6ede36819b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6773493
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>