drivers: serial: silabs: Don't fail to init if clock is on
It is not an error if the clock source for the UART is already
enabled. This may happen if a bootloader has used the UART and
not de-initialized it before booting the application.

(cherry picked from commit dadd657d89791ba75230b84eaf848b124a51817a)

Original-Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>
GitOrigin-RevId: dadd657d89791ba75230b84eaf848b124a51817a
Cr-Build-Id: 8712157887905019265
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8712157887905019265
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I34b8ddd4e9fbd55027ba03b92aef9814a6932221
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6643892
Commit-Queue: Jon Murphy <jpmurphy@google.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>