Revert "drivers: spi: mcux_flexcomm: use rxignore bit instead of dummy read."
This reverts commit a3530d6a43a199ce6a7df326038ea95431148d57.

This change incorrect if chip-select is via a GPIO pin,
released by spi_context_cs_control(), and not the controller
default chip-select pin that can be released using a final
FIFOWR.

When GPIO is used for chip-select control, the chip-select could
be released too soon because the transfer callback is invoked
when the last byte is put into the TX FIFO but that byte might
not yet have clocked out yet on the SPI pins. The rx part of
the transfer is really what completes the transfer so ignoring
it is incorrect.

We could try to special case and use ignore only when
spi_context_cs_control() does nothing, but since it is a very
small optimization, it doesn't seem worth the extra
complexity.

(cherry picked from commit 0c1e9c3e7a702acd92f83eb7223dea4f7c0ff93c)

Original-Signed-off-by: Mike J. Chen <mjchen@google.com>
GitOrigin-RevId: 0c1e9c3e7a702acd92f83eb7223dea4f7c0ff93c
Cr-Build-Id: 8706654121712404961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706654121712404961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iff6140c9f6a548e6dfb38ffdf7dc2b9d725b7a1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6845410
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Yuval Peress <peress@google.com>