Bluetooth: BAP: BA: Refactor discover to not do read
Refactor the bt_bap_broadcast_assistant_discover function
to not read receives at the end of discovery.
This makes the function more true to what it is supposed to
do, and significantly reduces the complexity of the
procedure and the read callback.

Users will be required to, if wanted, to read the
receive state themselves with the existing
bt_bap_broadcast_assistant_read_recv_state.

The start_offset had to be modified to avoid issues with
PA sync.

(cherry picked from commit f4e14dfb243fbefeb60842260b969cfcabd83390)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: f4e14dfb243fbefeb60842260b969cfcabd83390
Cr-Build-Id: 8693426963900677041
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8693426963900677041
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I52366ec634f18f9fa953bdaa39d818fcea8923a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7402168
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>