drivers: ethernet: vsc8541: Fixed inverted reset GPIO
For GPIOs driving active-low signals, such as the VSC8541's reset pin,
they are supposed to be declared as active low in the device tree, and
set to 1 to assert and 0 to clear. Change the driver as such so that it
does not leave the PHY stuck in reset when so configured.

Also changed all in-tree board DTS files for this PHY to properly
declare the reset GPIO as active low.

(cherry picked from commit 211e9ea6bd6d215a311289f408b5344bd782744f)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: 211e9ea6bd6d215a311289f408b5344bd782744f
Cr-Build-Id: 8708556658338805137
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8708556658338805137
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1e600c8a236c81f10ea4e78b610e1185968df6e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6781268
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>