drivers: flash_mspi_nor: Omit quad_enable_set() when QER is set to NONE
When the quad-enable-requirements property is set to "NONE" or is not
present, no Quad Enable operation should be performed.
This fixes an issue with the mx25uw6345g flash chip that is present
on the nRF54h20 DK and supports the Single I/O mode, but cannot be
used in that mode.

(cherry picked from commit 8415f7f7a75111dd85b84743e691799d68659e42)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: 8415f7f7a75111dd85b84743e691799d68659e42
Cr-Build-Id: 8711229266906812305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8711229266906812305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ie594480709fd591923387ae5958313a62450618e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6665512
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>