net: lwm2m: Make Bootstrap on Registration Failure configurable
The fallback to Bootstrap on a Registration Failure might not be desirable
for some applications. This change makes the default behavior configure
through Kconfig.

An alternative to configure "Bootstrap on Registration Failure" would be to
set the server object resource id 16 at runtime. However, this is tedious
and error-prone, as it will be reset when the server object is deleted and
might have to be set at different locations in the client code.

(cherry picked from commit 86dbcd2cf2bb42de532bce523d9b028414f02fb2)

Original-Signed-off-by: Marc Lasch <marc.lasch@husqvarnagroup.com>
GitOrigin-RevId: 86dbcd2cf2bb42de532bce523d9b028414f02fb2
Cr-Build-Id: 8710912179129064817
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8710912179129064817
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I18de219ed8091753d9044eafb897129376d0be28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6683753
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>