boards/raytac: Change Nordic license to Apache
For yaml files we do not normally even add a license, but if any it
should be Apache2 and not the Nordic one.
Note there is plenty of examples of yaml files very similar to this
in the Zephyr tree. So even if the author/submitter of the original
commit could have used as starting point a file from NCS, it would have
been the same if he would have used a Zephyr one.

(cherry picked from commit 789a0a8da7fe6d7786ff9746826fbfec4480f446)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 789a0a8da7fe6d7786ff9746826fbfec4480f446
Cr-Build-Id: 8710912179129064817
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8710912179129064817
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If77ebd40634f998c37aff830c575fdf14112e429
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6683756
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>