driver: i2c: npcx: remove unused member from struct i2c_ctrl_data
The member `is_port_set` in struct `i2c_ctrl_data` was added in
CL:6576538 to indicate whether the I2C port is initialized or not,
but it was not upstreamed. Since we decided to adopt the fix in
CL:6576800, CL:6581529 was pushed to remove the code that uses
`is_port_set`. As a result, `is_port_set` is now unused and should be
removed from the codebase entirely.

Change-Id: Idc8e5cf5e6b7ce0d0cd5044301faeda39e5612b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6665823
Commit-Queue: CH Lin (Jun) <chlin56@nuvoton.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CH Lin (Jun) <chlin56@nuvoton.com>