dts: nxp: Fix RT11xx DT files hierarchy
The DT hierachy of the RT11xx series was somewhat incoherent. The boards
targets were directly including the series level DTSI with no SOC dtsi
in between, and there existed an SOC DTSI that had to be separately
included by a different board file, which didn't include the series DTSI
itself. It seems that this was only working if you included the files
exactly in the correct order in specific board files. Also, as a result
of this change, need to (correctly) define the cpu core only in the DTSI
for that core, instead of in the series generic dtsi, because that DTSI
was actually written with incorrect syntax due to duplicated node labels
on nodes right next to each other in the same file, and was relying on
other DTSI files to delete the duplicate nodes in order for it to build.
So overall this was a mess, needed cleanup.

(cherry picked from commit dba69f3f14e5b84b411062f689d781360cefd934)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: dba69f3f14e5b84b411062f689d781360cefd934
Cr-Build-Id: 8710527140294911377
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8710527140294911377
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I262f4946d2b1547f3052b457e7fe443651c48892
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6686837
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>