drivers: serial: nrfx_uarte: Add mode with TIMER byte counting
Add mode to be used on UARTE with frame timeout which is using a bounce
buffers and TIMER to count bytes. This mode shall be used to reliably
receive data without HWFC as frame timeout approach is not 100% reliable
because it can loose or corrupt a byte when new byte arrives after
frame timeout is detected but before it is fully handled. This mode is
similar to the one enabled with CONFIG_UART_x_NRF_HW_ASYNC but
additional bounce buffers are used and UARTE is receiving data to
internal buffers and copies data to the user buffer. Legacy apporach
cannot be used because in new SoC DMA attempts to copy data in words
so when byte is received it stays in the DMA internal buffer until
4 bytes are received or end of transfer happens then internal DMA
buffer is flushed.

(cherry picked from commit f1f0b03aeaa9d54ed413196b39e3fd5034bd2efb)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: f1f0b03aeaa9d54ed413196b39e3fd5034bd2efb
Cr-Build-Id: 8698221080423948257
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8698221080423948257
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I9e0ce205653a3e1b7b4436a31947281ac3444428
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7158249
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>