tests: spi_loopback: skip hold_on_cs test if not supported
This early exit section is only reached if there is an error.
In the case of not supported, the ret variable will be 0 and zassert
won't happen. This is because the test was meant to be skipped, not
passed or failed, in this case. So add the skip call.

(cherry picked from commit 0574ac7e845876c8e9fee0c8f3974d519a7ce086)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 0574ac7e845876c8e9fee0c8f3974d519a7ce086
Cr-Build-Id: 8710504492956259633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8710504492956259633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ifaec91ef5e3cba00a5c9afb951be46f67e17534b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6698069
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>