scripts: ci: differentiate warnings from failures in compliance_check
We have a few checks that report warnings, and for some of them it's
actually the _only_ type of failure they report, so it can be confusing
when the output of check_compliance.py says the "check failed" and lists
all the things that are effectively warnings as "ERROR"s.
This commit makes the output more clear by differentiating between
warnings and actual errors, both when printing out individual failures,
but also when printing the summary for each test case (ie. a test case
is said to be "failing" only when it has at least one error, and is only
showing up as having warnings if it only contains... warnings).

(cherry picked from commit 5a1065c37f8864296923e590e557533a5b3aa241)

Original-Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>
GitOrigin-RevId: 5a1065c37f8864296923e590e557533a5b3aa241
Cr-Build-Id: 8705566957733089633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8705566957733089633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I60b192e321140c700a8a6c2432b7223b745360f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6879923
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>