dts: arm: st: stm32c0: remove U suffix from "clocks" in DTSI
PR 79683 added unnecessary U suffix to constants in DTSI.
This bloats the files for no reason and is starting to spread as people
use the DTSI for reference in other places, and so copy this bad pattern.

Remove the useless U suffixes in DTSI files for this series.

(cherry picked from commit bc541f293a305d34fcfa3a75103d003321e590bc)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain@st.com>
GitOrigin-RevId: bc541f293a305d34fcfa3a75103d003321e590bc
Cr-Build-Id: 8704185354746732337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704185354746732337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2ff64f46512abb7aea34b20bb165bf29bcc10ae6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6932175
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>