dts: arm: st: stm32f2: remove U suffix from "clocks" in DTSI
PR 79683 added unnecessary U suffix to constants in DTSI.
This bloats the files for no reason and is starting to spread as people
use the DTSI for reference in other places, and so copy this bad pattern.

Remove the useless U suffixes in DTSI files for this series.

(cherry picked from commit 094c2258b8660d5bd05fc2265f51902ded7d8ebd)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain@st.com>
GitOrigin-RevId: 094c2258b8660d5bd05fc2265f51902ded7d8ebd
Cr-Build-Id: 8704185354746732337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704185354746732337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I717da802731c2bc6d07d27dba231a6ac1ee792c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6932178
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>