dts: arm: st: stm32h7: remove U suffix from "clocks" in DTSI
PR 79683 added unnecessary U suffix to constants in DTSI.
This bloats the files for no reason and is starting to spread as people
use the DTSI for reference in other places, and so copy this bad pattern.

Remove the useless U suffixes in DTSI files for this series.

(cherry picked from commit 0722161a50faf1e8bcea01d8465e86312eef3b4b)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain@st.com>
GitOrigin-RevId: 0722161a50faf1e8bcea01d8465e86312eef3b4b
Cr-Build-Id: 8704185354746732337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704185354746732337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic2da8c29b70984340d93db0bc526bca42e61881e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6932185
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>