dts: arm: st: stm32l1: remove U suffix from "clocks" in DTSI
PR 79683 added unnecessary U suffix to constants in DTSI.
This bloats the files for no reason and is starting to spread as people
use the DTSI for reference in other places, and so copy this bad pattern.

Remove the useless U suffixes in DTSI files for this series.

(cherry picked from commit 0744927aebefd3b1676be7cc880c9476017a1a32)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain@st.com>
GitOrigin-RevId: 0744927aebefd3b1676be7cc880c9476017a1a32
Cr-Build-Id: 8704185354746732337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704185354746732337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I84d8dac4d3c07493d5bb250c0b50380cf70bef93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6932188
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>