dts: arm: st: stm32wl: remove U suffix from "clocks" in DTSI
PR 79683 added unnecessary U suffix to constants in DTSI.
This bloats the files for no reason and is starting to spread as people
use the DTSI for reference in other places, and so copy this bad pattern.

Remove the useless U suffixes in DTSI files for this series.

(cherry picked from commit fb55091fd90cabf350789b1356128269c07e0e86)

Original-Signed-off-by: Mathieu Choplain <mathieu.choplain@st.com>
GitOrigin-RevId: fb55091fd90cabf350789b1356128269c07e0e86
Cr-Build-Id: 8704185354746732337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704185354746732337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5333eb8634ad3b6fb748d6ecc5d5c41c344556f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6932196
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>