samples: tfm: tfm_psa_test: Add temporary isolation workaround
Add a workaround which sets the isolation level to 2 in CMake.
The TF-M tests require the large TF-M profile because it supports
the full list of crypto algorithms needed, not because of the
solation level.

For the TF-M tests the isolation level is irrelevant so we
set it to 2 here so that we don't exclude the platforms which
don't support the isolation level 3.

I communicated with a TF-M maintainer and he informed me that this
workaround will be included in the TF-M 2.2 release branch later
so this is a short lived workaround that can be reverted when the
relevant commit from the TF-M branch is cherry-picked.

(cherry picked from commit 0081207078a908324e427b98868464123b0b0f82)

Original-Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>
GitOrigin-RevId: 0081207078a908324e427b98868464123b0b0f82
Cr-Build-Id: 8709621173687530433
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8709621173687530433
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iff8fbfb6e5f9dabfa06a2138327f884c541f6fa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6729621
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>