tests: spi_loopback: nxp_lpspi: Increase latency for DMA
The DMA-based LPSPI is inherently going to have a high latency due to
the requirement for setting up the DMA transfer and the DMA driver may
be extremely bloated, this is no fault of the spi driver and not
relevant to the test. The driver we want to test for latency is the
CPU-based one, the DMA-based one is more prioritizing bandwidth over
latency.

(cherry picked from commit 6f26597ddb750f12555b2df2cc594dc01acd7278)

Original-Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
GitOrigin-RevId: 6f26597ddb750f12555b2df2cc594dc01acd7278
Cr-Build-Id: 8707333598751242721
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707333598751242721
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ibd559fefebe9cee4ec37df1f1cda5d43ce4b83a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6822672
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>