scripts/logging: Add support for numbered %ll and %hh
Python does not support values of long-long (ll) or short-short (hh) in
string-formatting. The current parser is correctly removing the first "h"
or "l" from the format, to make it possible to correctly parse these
strings when they are in the dictionary.

However, the current implementation does not correctly handle the case of
having a number before the "ll" or "hh" argument - which is supported by
both C and Python.

This commit updates the parser to correctly handle these cases, by changing
the simple replace operator with a regex replace, which is able to
correctly handle this case.

Before this commit, the string of:
    "This is my long variable: %08llX"
will stay the same, which will throw a ValueError when parsing this
dictionary message:
    "ValueError: unsupported format character 'h' (0x68) at index ".
After this commit, the string will be correctly fixed to
    "This is my long variable: %08lX"
which is parsed correctly.

(cherry picked from commit cb94b1ba2a4d13c1d62537c8ab1be88ef78e2296)

Original-Signed-off-by: Omri Sarig <omsi@demant.com>
GitOrigin-RevId: cb94b1ba2a4d13c1d62537c8ab1be88ef78e2296
Cr-Build-Id: 8706112713020869521
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8706112713020869521
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic28befa32e3031309c8d6fa22b72f59a2c9ca12e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6861885
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>