scripts/logging: Add support to RTT reading from JLink
Update the live logger to be able to read the logs from JLink, over the RTT
protocol.
The support for the jlink-rtt mode is done similarly to the currently
existing support for files and serials.

The implementation for opening the jlink for reading was inspired by the
available example from the pylink python implementation here:

In contrast to the other readers, the jlink-rtt does not support being
polled from 'select'.
To make it possible to work with such readers, the code was updated to
allow for busy-waits, instead of smart polling.
This change does not affect the current readers, but support the new
reader, and allows for future readers which does not support smart polling.

(cherry picked from commit f55dcc91dfa7d7963f1c1ddb6356390057ca13e9)

Original-https: //github.com/square/pylink/blob/master/examples/pylink-rtt
Original-Signed-off-by: Omri Sarig <omsi@demant.com>
GitOrigin-RevId: f55dcc91dfa7d7963f1c1ddb6356390057ca13e9
Cr-Build-Id: 8704208004692697937
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704208004692697937
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I9f0ee28327d823c47c764a1dc3c87cd2ee0bb4c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6929272
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>