tracing: ctf: take IRQ lock before generating timestamp
CTF tracing relies on all packet timestamps being increasing- if a
timestamp of a later packet is less than the prior packet, the CTF
parser assumes the timestamp field has overflowed and adds to the
overall timestamp to account for this.

When taking a timestamp for a CTF packet, it was possible for an interrupt
to occur before the packet was submitted to the tracing core framework
but after the timestamp was generated. The interrupt itself would
generate a tracing event with a later timestamp then the packet in
question, leading to the packets being recorded out of order.

To resolve this, take an IRQ lock before generating the timestamp and
release it after submitting the packet to the tracing core.

(cherry picked from commit c7e60346a03eec8604c46c3b8bc98d22b6e8216f)

Original-Signed-off-by: Daniel DeGrasse <ddegrasse@tenstorrent.com>
GitOrigin-RevId: c7e60346a03eec8604c46c3b8bc98d22b6e8216f
Cr-Build-Id: 8704253302701214961
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8704253302701214961
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I7378713e6d87e03285528ab1a9a774c28cd35565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6927721
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>